refactor: extract formatKeyStroke to common
